+++ app/src/Provider.tsx
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 import { CssBaseline, StyledEngineProvider, ThemeProvider } from '@mui/material';
-import { AuthProvider } from '@opentalk/react-redux-appauth';
+import { AuthProvider } from '@opentalk/redux-oidc';
 import React from 'react';
 
 import { createOpenTalkTheme } from './assets/themes/opentalk';
@@ -11,8 +11,7 @@ import { useAppSelector } from './hooks';
 import BreakoutRoomProvider from './provider/BreakoutRoomProvider';
 import FullscreenProvider from './provider/FullscreenProvider';
 import SnackbarProvider from './provider/SnackbarProvider';
-import store from './store';
-import { selectBaseUrl, selectOidcConfig } from './store/slices/configSlice';
+import { selectBaseUrl, selectControllerUrl, selectOidcConfig } from './store/slices/configSlice';
 import { ConnectionState, selectRoomConnectionState } from './store/slices/roomSlice';
 
 interface ProviderProps {
@@ -24,21 +23,19 @@ const Provider = ({ children }: ProviderProps) => {
   const baseUrl = useAppSelector(selectBaseUrl);
   const roomState = useAppSelector(selectRoomConnectionState);
   const inRoom = roomState === ConnectionState.Online || roomState === ConnectionState.Leaving;
+  const controllerBasedUrl = useAppSelector(selectControllerUrl);
 
   return (
     <StyledEngineProvider injectFirst>
       <AuthProvider
-        store={store}
-        authority={oidcConfig.authority}
-        clientId={oidcConfig.clientId}
-        signOutRedirectUri={new URL(oidcConfig.signOutRedirectUri, baseUrl).toString()}
-        redirectUri={new URL(oidcConfig.redirectPath, baseUrl).toString()}
-        // We can use the popup_redirect_uri at this point to just share a single route.
-        // The iframe content will never be visible.
-        silentRedirectUri={new URL(oidcConfig.popupRedirectPath, baseUrl).toString()}
-        popupRedirectUri={new URL(oidcConfig.popupRedirectPath, baseUrl).toString()}
-        scope={oidcConfig.scope}
-        silentSignin={true}
+        configuration={{
+          authority: oidcConfig.authority,
+          clientId: oidcConfig.clientId,
+          redirectUri: new URL(oidcConfig.redirectPath, baseUrl).toString(),
+          scope: oidcConfig.scope,
+          baseUrl: controllerBasedUrl,
+          signOutRedirectUri: new URL(oidcConfig.signOutRedirectUri, baseUrl).toString(),
+        }}
       >
         <ThemeProvider theme={inRoom ? createOpenTalkTheme('dark') : createOpenTalkTheme()}>
           <CssBaseline />