+++ packages/redux-oidc/src/store/index.ts
@@ -1,0 +1,7 @@
+import { authSlice, AuthState } from './authSlice';
+import authReducer from './authSlice';
+
+export * from './authActions';
+export * from './authSlice';
+export type { AuthState };
+export { authSlice, authReducer };